+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/FormatControl.cs
@@ -489,794 +489,6 @@ internal partial class FormatControl : UserControl
         public abstract void PushFormatStringIntoFormatType(string formatString);
     }
 
-    private class NoFormattingFormatType : FormatTypeClass
-    {
-        public override string TopLabelString
-        {
-            get
-            {
-                return SR.BindingFormattingDialogFormatTypeNoFormattingExplanation;
-            }
-        }
-
-        public override string SampleString
-        {
-            get
-            {
-                return "-1234.5";
-            }
-        }
-
-        public override bool DropDownVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool ListBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool FormatLabelVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override string FormatString
-        {
-            get
-            {
-                return "";
-            }
-        }
-
-        public override bool FormatStringTextBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool Parse(string formatString)
-        {
-            return false;
-        }
-
-        public override void PushFormatStringIntoFormatType(string formatString)
-        {
-            // nothing to do;
-        }
-
-        public override string ToString()
-        {
-            return SR.BindingFormattingDialogFormatTypeNoFormatting;
-        }
-    }
-
-    private class NumericFormatType : FormatTypeClass
-    {
-        private readonly FormatControl _owner;
-
-        public NumericFormatType(FormatControl owner)
-        {
-            _owner = owner;
-        }
-
-        public override string TopLabelString
-        {
-            get
-            {
-                return SR.BindingFormattingDialogFormatTypeNumericExplanation;
-            }
-        }
-
-        public override string SampleString
-        {
-            get
-            {
-                return (-1234.5678).ToString(FormatString, CultureInfo.CurrentCulture);
-            }
-        }
-
-        public override bool DropDownVisible
-        {
-            get
-            {
-                return true;
-            }
-        }
-
-        public override bool ListBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool FormatLabelVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override string FormatString
-        {
-            get
-            {
-                switch ((int)_owner.decimalPlacesUpDown.Value)
-                {
-                    case 0:
-                        return "N0";
-                    case 1:
-                        return "N1";
-                    case 2:
-                        return "N2";
-                    case 3:
-                        return "N3";
-                    case 4:
-                        return "N4";
-                    case 5:
-                        return "N5";
-                    case 6:
-                        return "N6";
-                    default:
-                        Debug.Fail("decimalPlacesUpDown should allow only up to 6 digits");
-                        return "";
-                }
-            }
-        }
-
-        public override bool FormatStringTextBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public static bool ParseStatic(string formatString)
-        {
-            return formatString.Equals("N0") ||
-                   formatString.Equals("N1") ||
-                   formatString.Equals("N2") ||
-                   formatString.Equals("N3") ||
-                   formatString.Equals("N4") ||
-                   formatString.Equals("N5") ||
-                   formatString.Equals("N6");
-        }
-
-        public override bool Parse(string formatString)
-        {
-            return ParseStatic(formatString);
-        }
-
-        public override void PushFormatStringIntoFormatType(string formatString)
-        {
-#if DEBUG
-            Debug.Assert(Parse(formatString), "we only push valid strings");
-#endif // DEBUG
-            if (formatString.Equals("N0"))
-            {
-                _owner.decimalPlacesUpDown.Value = 0;
-            }
-            else if (formatString.Equals("N1"))
-            {
-                _owner.decimalPlacesUpDown.Value = 1;
-            }
-            else if (formatString.Equals("N2"))
-            {
-                _owner.decimalPlacesUpDown.Value = 2;
-            }
-            else if (formatString.Equals("N3"))
-            {
-                _owner.decimalPlacesUpDown.Value = 3;
-            }
-            else if (formatString.Equals("N4"))
-            {
-                _owner.decimalPlacesUpDown.Value = 4;
-            }
-            else if (formatString.Equals("N5"))
-            {
-                _owner.decimalPlacesUpDown.Value = 5;
-            }
-            else if (formatString.Equals("N6"))
-            {
-                _owner.decimalPlacesUpDown.Value = 6;
-            }
-        }
-
-        public override string ToString()
-        {
-            return SR.BindingFormattingDialogFormatTypeNumeric;
-        }
-    }
-
-    private class CurrencyFormatType : FormatTypeClass
-    {
-        private readonly FormatControl _owner;
-
-        public CurrencyFormatType(FormatControl owner)
-        {
-            _owner = owner;
-        }
-
-        public override string TopLabelString
-        {
-            get
-            {
-                return SR.BindingFormattingDialogFormatTypeCurrencyExplanation;
-            }
-        }
-
-        public override string SampleString
-        {
-            get
-            {
-                return (-1234.5678).ToString(FormatString, CultureInfo.CurrentCulture);
-            }
-        }
-
-        public override bool DropDownVisible
-        {
-            get
-            {
-                return true;
-            }
-        }
-
-        public override bool ListBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool FormatLabelVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override string FormatString
-        {
-            get
-            {
-                switch ((int)_owner.decimalPlacesUpDown.Value)
-                {
-                    case 0:
-                        return "C0";
-                    case 1:
-                        return "C1";
-                    case 2:
-                        return "C2";
-                    case 3:
-                        return "C3";
-                    case 4:
-                        return "C4";
-                    case 5:
-                        return "C5";
-                    case 6:
-                        return "C6";
-                    default:
-                        Debug.Fail("decimalPlacesUpDown should allow only up to 6 digits");
-                        return "";
-                }
-            }
-        }
-
-        public override bool FormatStringTextBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public static bool ParseStatic(string formatString)
-        {
-            return formatString.Equals("C0") ||
-                   formatString.Equals("C1") ||
-                   formatString.Equals("C2") ||
-                   formatString.Equals("C3") ||
-                   formatString.Equals("C4") ||
-                   formatString.Equals("C5") ||
-                   formatString.Equals("C6");
-        }
-
-        public override bool Parse(string formatString)
-        {
-            return ParseStatic(formatString);
-        }
-
-        public override void PushFormatStringIntoFormatType(string formatString)
-        {
-#if DEBUG
-            Debug.Assert(Parse(formatString), "we only push valid strings");
-#endif // DEBUG
-            if (formatString.Equals("C0"))
-            {
-                _owner.decimalPlacesUpDown.Value = 0;
-            }
-            else if (formatString.Equals("C1"))
-            {
-                _owner.decimalPlacesUpDown.Value = 1;
-            }
-            else if (formatString.Equals("C2"))
-            {
-                _owner.decimalPlacesUpDown.Value = 2;
-            }
-            else if (formatString.Equals("C3"))
-            {
-                _owner.decimalPlacesUpDown.Value = 3;
-            }
-            else if (formatString.Equals("C4"))
-            {
-                _owner.decimalPlacesUpDown.Value = 4;
-            }
-            else if (formatString.Equals("C5"))
-            {
-                _owner.decimalPlacesUpDown.Value = 5;
-            }
-            else if (formatString.Equals("C6"))
-            {
-                _owner.decimalPlacesUpDown.Value = 6;
-            }
-        }
-
-        public override string ToString()
-        {
-            return SR.BindingFormattingDialogFormatTypeCurrency;
-        }
-    }
-
-    private class DateTimeFormatType : FormatTypeClass
-    {
-        private readonly FormatControl _owner;
-
-        public DateTimeFormatType(FormatControl owner)
-        {
-            _owner = owner;
-        }
-
-        public override string TopLabelString
-        {
-            get
-            {
-                return SR.BindingFormattingDialogFormatTypeDateTimeExplanation;
-            }
-        }
-
-        public override string SampleString
-        {
-            get
-            {
-                if (_owner.dateTimeFormatsListBox.SelectedItem is null)
-                {
-                    return "";
-                }
-
-                return s_dateTimeFormatValue.ToString(FormatString, CultureInfo.CurrentCulture);
-            }
-        }
-
-        public override bool DropDownVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool ListBoxVisible
-        {
-            get
-            {
-                return true;
-            }
-        }
-
-        public override bool FormatLabelVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override string FormatString
-        {
-            get
-            {
-                DateTimeFormatsListBoxItem item = _owner.dateTimeFormatsListBox.SelectedItem as DateTimeFormatsListBoxItem;
-                return item.FormatString;
-            }
-        }
-
-        public override bool FormatStringTextBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public static bool ParseStatic(string formatString)
-        {
-            return formatString.Equals("d") ||
-                   formatString.Equals("D") ||
-                   formatString.Equals("f") ||
-                   formatString.Equals("F") ||
-                   formatString.Equals("g") ||
-                   formatString.Equals("G") ||
-                   formatString.Equals("t") ||
-                   formatString.Equals("T") ||
-                   formatString.Equals("M");
-        }
-
-        public override bool Parse(string formatString)
-        {
-            return ParseStatic(formatString);
-        }
-
-        public override void PushFormatStringIntoFormatType(string formatString)
-        {
-#if DEBUG
-            Debug.Assert(Parse(formatString), "we only push valid strings");
-#endif // DEBUG
-            int selectedIndex = -1;
-
-            if (formatString.Equals("d"))
-            {
-                selectedIndex = 0;
-            }
-            else if (formatString.Equals("D"))
-            {
-                selectedIndex = 1;
-            }
-            else if (formatString.Equals("f"))
-            {
-                selectedIndex = 2;
-            }
-            else if (formatString.Equals("F"))
-            {
-                selectedIndex = 3;
-            }
-            else if (formatString.Equals("g"))
-            {
-                selectedIndex = 4;
-            }
-            else if (formatString.Equals("G"))
-            {
-                selectedIndex = 5;
-            }
-            else if (formatString.Equals("t"))
-            {
-                selectedIndex = 6;
-            }
-            else if (formatString.Equals("T"))
-            {
-                selectedIndex = 7;
-            }
-            else if (formatString.Equals("M"))
-            {
-                selectedIndex = 8;
-            }
-
-            _owner.dateTimeFormatsListBox.SelectedIndex = selectedIndex;
-        }
-
-        public override string ToString()
-        {
-            return SR.BindingFormattingDialogFormatTypeDateTime;
-        }
-    }
-
-    private class ScientificFormatType : FormatTypeClass
-    {
-        private readonly FormatControl _owner;
-
-        public ScientificFormatType(FormatControl owner)
-        {
-            _owner = owner;
-        }
-
-        public override string TopLabelString
-        {
-            get
-            {
-                return SR.BindingFormattingDialogFormatTypeScientificExplanation;
-            }
-        }
-
-        public override string SampleString
-        {
-            get
-            {
-                return (-1234.5678).ToString(FormatString, CultureInfo.CurrentCulture);
-            }
-        }
-
-        public override bool DropDownVisible
-        {
-            get
-            {
-                return true;
-            }
-        }
-
-        public override bool ListBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool FormatLabelVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override string FormatString
-        {
-            get
-            {
-                switch ((int)_owner.decimalPlacesUpDown.Value)
-                {
-                    case 0:
-                        return "E0";
-                    case 1:
-                        return "E1";
-                    case 2:
-                        return "E2";
-                    case 3:
-                        return "E3";
-                    case 4:
-                        return "E4";
-                    case 5:
-                        return "E5";
-                    case 6:
-                        return "E6";
-                    default:
-                        Debug.Fail("decimalPlacesUpDown should allow only up to 6 digits");
-                        return "";
-                }
-            }
-        }
-
-        public override bool FormatStringTextBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public static bool ParseStatic(string formatString)
-        {
-            return formatString.Equals("E0") ||
-                   formatString.Equals("E1") ||
-                   formatString.Equals("E2") ||
-                   formatString.Equals("E3") ||
-                   formatString.Equals("E4") ||
-                   formatString.Equals("E5") ||
-                   formatString.Equals("E6");
-        }
-
-        public override bool Parse(string formatString)
-        {
-            return ParseStatic(formatString);
-        }
-
-        public override void PushFormatStringIntoFormatType(string formatString)
-        {
-#if DEBUG
-            Debug.Assert(Parse(formatString), "we only push valid strings");
-#endif // DEBUG
-            if (formatString.Equals("E0"))
-            {
-                _owner.decimalPlacesUpDown.Value = 0;
-            }
-            else if (formatString.Equals("E1"))
-            {
-                _owner.decimalPlacesUpDown.Value = 1;
-            }
-            else if (formatString.Equals("E2"))
-            {
-                _owner.decimalPlacesUpDown.Value = 2;
-            }
-            else if (formatString.Equals("E3"))
-            {
-                _owner.decimalPlacesUpDown.Value = 3;
-            }
-            else if (formatString.Equals("E4"))
-            {
-                _owner.decimalPlacesUpDown.Value = 4;
-            }
-            else if (formatString.Equals("E5"))
-            {
-                _owner.decimalPlacesUpDown.Value = 5;
-            }
-            else if (formatString.Equals("E6"))
-            {
-                _owner.decimalPlacesUpDown.Value = 6;
-            }
-        }
-
-        public override string ToString()
-        {
-            return SR.BindingFormattingDialogFormatTypeScientific;
-        }
-    }
-
-    private class CustomFormatType : FormatTypeClass
-    {
-        private readonly FormatControl _owner;
-
-        public CustomFormatType(FormatControl owner)
-        {
-            _owner = owner;
-        }
-
-        public override string TopLabelString
-        {
-            get
-            {
-                return SR.BindingFormattingDialogFormatTypeCustomExplanation;
-            }
-        }
-
-        public override string SampleString
-        {
-            get
-            {
-                string formatString = FormatString;
-
-                if (string.IsNullOrEmpty(formatString))
-                {
-                    return "";
-                }
-
-                string label = "";
-
-                // first see if the formatString is one of DateTime's format strings
-                if (DateTimeFormatType.ParseStatic(formatString))
-                {
-                    label = s_dateTimeFormatValue.ToString(formatString, CultureInfo.CurrentCulture);
-                }
-
-                // the formatString was not one of DateTime's strings
-                // Try a double
-                if (label.Equals(""))
-                {
-                    try
-                    {
-                        label = (-1234.5678).ToString(formatString, CultureInfo.CurrentCulture);
-                    }
-                    catch (FormatException)
-                    {
-                        label = "";
-                    }
-                }
-
-                // double failed.
-                // Try an Int
-                if (label.Equals(""))
-                {
-                    try
-                    {
-                        label = (-1234).ToString(formatString, CultureInfo.CurrentCulture);
-                    }
-                    catch (FormatException)
-                    {
-                        label = "";
-                    }
-                }
-
-                // int failed.
-                // apply the formatString to the dateTime value
-                if (label.Equals(""))
-                {
-                    try
-                    {
-                        label = s_dateTimeFormatValue.ToString(formatString, CultureInfo.CurrentCulture);
-                    }
-                    catch (FormatException)
-                    {
-                        label = "";
-                    }
-                }
-
-                if (label.Equals(""))
-                {
-                    label = SR.BindingFormattingDialogFormatTypeCustomInvalidFormat;
-                }
-
-                return label;
-            }
-        }
-
-        public override bool DropDownVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool ListBoxVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override bool FormatStringTextBoxVisible
-        {
-            get
-            {
-                return true;
-            }
-        }
-
-        public override bool FormatLabelVisible
-        {
-            get
-            {
-                return false;
-            }
-        }
-
-        public override string FormatString
-        {
-            get
-            {
-                return _owner.customStringTextBox.Text;
-            }
-        }
-
-        public static bool ParseStatic(string formatString)
-        {
-            // anything goes...
-            return true;
-        }
-
-        public override bool Parse(string formatString)
-        {
-            return ParseStatic(formatString);
-        }
-
-        public override void PushFormatStringIntoFormatType(string formatString)
-        {
-            _owner.customStringTextBox.Text = formatString;
-        }
-
-        public override string ToString()
-        {
-            return SR.BindingFormattingDialogFormatTypeCustom;
-        }
-    }
-
     private void nullValueTextBox_TextChanged(object sender, EventArgs e)
     {
         Dirty = true;
